Refactored dateFormatter and fixed how it's working.
