+++ public/javascripts/task-comments.js
@@ -1,4 +1,4 @@
-import { handleErrors } from "./utils.js";
+import { handleErrors, dateFormatter } from "./utils.js";
 import { fetchTasks } from './app.js';
 
 
@@ -30,75 +30,10 @@ export const fetchTask = async (taskId) => {
         return;
     }
 
-
-    const today = new Date().toJSON().slice(0,10)
-
-    let tomorrow = today.split('-');
-    let month = tomorrow[tomorrow.length - 2]
-    let day = tomorrow[tomorrow.length - 1]
-
-    // TODO - make sure new month and new day are in the correct format
-
-    if (day === '31' &&
-        (month === '01' ||
-        month === '03' ||
-        month === '05' ||
-        month === '07' ||
-        month === '08' ||
-        month === '10' ||
-        month === '12')) {
-            day = '01';
-            if (month !== '12' && Number(month) >= 9) {
-                month = (Number(month) + 1).toString()
-            } else if (month === '12') {
-                month = '01';
-            }
-            else {
-                month = '0' + (Number(month) + 1).toString();
-            }
-    } else if (
-        day === '30' && (
-        month === '02' ||
-        month === '04' ||
-        month === '06' ||
-        month === '09' ||
-        month === '11' ))
-        {
-        day = '01';
-        if (Number(month) <= 9) {
-            month = (Number(month) + 1).toString()
-        } else {
-            month = '0' + (Number(month) + 1).toString();
-        }
-    } else {
-        if (Number(day) >= 9) {
-            day = (Number(day) + 1).toString();
-        } else {
-            day = '0' + (Number(day) + 1).toString();
-        }
-    }
-
-    tomorrow[tomorrow.length - 1] = day;
-    tomorrow[tomorrow.length - 2] = month;
-
-    tomorrow = tomorrow.join('-');
-
     const { task } = await res.json();
 
-    let due = task.dueDate.slice(0,10);
 
-    if (due === today) {
-        due = 'Today';
-    }
-
-    if (due === tomorrow) {
-        due = 'Tomorrow';
-    }
-
-    if (Number(due.slice(9)) < Number(today.slice(9))) {
-        due = 'OVERDUE!'
-    }
-
+    const due = dateFormatter(task);
 
     const taskInfo = document.querySelector('.fiona');
     if (!task.givenTo) task.givenTo = '';
@@ -151,9 +86,6 @@ export const fetchTask = async (taskId) => {
         deleteTask(task.id);
         const taskInfo = document.querySelector(`.task-${task.id}`);
         taskInfo.hidden = true;
-
-        // const taskOnList = document.querySelector(`.task-info-${task.id}`);
-        // taskOnList.remove();
     })
 
 }
